requirements.txt
# Sprint Agent API Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.40
langchain-core==0.1.52
openai==1.3.8
PyGithub==1.59.1
atlassian-python-api==3.41.5
jira==3.8.0
python-dotenv==1.0.0
typing-extensions==4.7.1

# Additional dependencies for the jira_agent module
pydantic==2.8.2
requests==2.31.0
pydantic-core==2.20.1
